# internal/http/assets/openapi.yaml
openapi: 3.0.3
info:
  title: Probelet API
  version: 0.1.0
  license:
    name: MIT
    url: https://github.com/Alkindi42/probelet/blob/main/LICENSE
  description: |
    Probelet is a lightweight HTTP service for simulating application and system
    behaviors in test, staging, and platform environments.

    It is designed for SREs, platform engineers, and developers who need to:
      - test liveness and readiness probes
      - simulate delayed or failing upstreams
      - inspect incoming HTTP requests
      - trigger bounded CPU and memory stress

    Probelet is intentionally:
      - stateless
      - deterministic
      - safe by default
      - easy to run as a single binary or container

    Authentication:
      - public discovery and probe endpoints can remain unauthenticated
      - protected endpoints require `X-Probelet-Token` when `PROBELET_TOKEN` is set
servers:
  - url: http://localhost:8000
    description: Local Probelet instance
  - url: /
    description: Relative base URL when documentation is served by Probelet itself
security: []
...
components:
  securitySchemes:
    ProbeletToken:
      type: apiKey
      in: header
      name: X-Probelet-Token
      description: |
        Static token authentication for protected Probelet endpoints.

        This header is required only when Probelet is started with the
        `PROBELET_TOKEN` environment variable set.
...
  headers:
    XRequestId:
      description: |
        Request correlation identifier.

        Probelet generates and returns this header on all endpoints except `/echo`.

        On `/echo`, the header is only returned if it was already present on the
        incoming request.
      schema:
        type: string
    CacheControl:
      description: Cache policy for the response.
      schema:
        type: string
        example: no-store
    XContentTypeOptions:
      description: Prevents MIME type sniffing.
      schema:
        type: string
        example: nosniff
  schemas:
    ErrorResponse:
      type: object
      required:
        - ok
        - message
      properties:
        ok:
          type: boolean
          enum: [false]
        message:
          type: string
      additionalProperties: false
...
    HeaderValues:
      type: object
      additionalProperties:
        type: array
        items:
          type: string
    QueryValues:
      type: object
      additionalProperties:
        type: array
        items:
          type: string
    EchoBody:
      type: object
      required:
        - content
        - bytes
        - is_truncated
      properties:
        content:
          type: string
          description: |
            Request body content as UTF-8 text when possible.
            If the body is not valid UTF-8, the content is returned as a base64-encoded string.
        bytes:
          type: integer
          format: int64
          description: Number of bytes returned in the echoed body payload, after truncation if applicable.
        is_truncated:
          type: boolean
      additionalProperties: false
    EchoData:
      type: object
      required:
        - body
        - client_ip
        - content_length
        - content_type
        - forwarded_for
        - headers
        - host
        - method
        - path
        - proto
        - query
        - raw_query
        - remote_addr
        - user_agent
        - x_real_ip
      properties:
        body:
          $ref: '#/components/schemas/EchoBody'
        client_ip:
          type: string
        content_length:
          type: integer
          format: int64
        content_type:
          type: string
        forwarded_for:
          type: string
        headers:
          $ref: '#/components/schemas/HeaderValues'
        host:
          type: string
        method:
          type: string
        path:
          type: string
        proto:
          type: string
        query:
          $ref: '#/components/schemas/QueryValues'
        raw_query:
          type: string
        remote_addr:
          type: string
        user_agent:
          type: string
        x_real_ip:
          type: string
      additionalProperties: false
    EchoResponse:
      type: object
      required:
        - ok
        - message
        - data
      properties:
        ok:
          type: boolean
          enum: [true]
        message:
          type: string
          enum: [echo]
        data:
          $ref: '#/components/schemas/EchoData'
      additionalProperties: false
...
paths:
...
  /echo:
    get:
      tags: [debugging]
      summary: Echo request details
      description: |
        Returns a structured representation of the incoming request.

        This endpoint is useful for inspecting:
          - headers
          - query parameters
          - request body
          - client address and forwarded headers
          - content type and protocol information

        The current implementation accepts any HTTP method.

        GET and POST are documented here as representative examples.

        When Probelet is started with `PROBELET_TOKEN` set, this endpoint requires
        the `X-Probelet-Token` header.
      operationId: getEcho
      security:
        - ProbeletToken: []
      responses:
        '200':
          description: Request details echoed back.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EchoResponse'
        '400':
          description: Failed to read request body.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid authentication token.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [debugging]
      summary: Echo request details with body

      description: |
        Returns a structured representation of the incoming request.

        This endpoint accepts any HTTP method. POST is documented here as a
        representative example for requests carrying a body.

        The request body is read and returned in a bounded representation.

        When Probelet is started with `PROBELET_TOKEN` set, this endpoint requires
        the `X-Probelet-Token` header.
      operationId: postEcho
      security:
        - ProbeletToken: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
          text/plain:
            schema:
              type: string
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Request details echoed back.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EchoResponse'
        '400':
          description: Failed to read request body.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid authentication token.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...